MandelBlockFilling(): Two setups

+++ t02_multithreading-mandelbrot/MandelBlockFilling.h
@@ -28,7 +28,7 @@
 
 #include "Processing.h"
 
-typedef float MbVal;
+typedef double MbVal;
 
 struct ConfigMandelbrot
 {
